Role

+++ dotnet_project/Areas/Admin/Controllers/ProductController.cs
@@ -85,7 +85,7 @@ namespace dotnet_project.Areas.Admin.Controllers
             return View(product);
         }
 
-        public async Task<IActionResult> Edit(int Id)
+        public async Task<IActionResult> Edit(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             ViewBag.Categories = new SelectList(_dataContext.Categories, "Id", "Name", product.CategoryId);
@@ -96,7 +96,7 @@ namespace dotnet_project.Areas.Admin.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int Id, ProductModel product)
+        public async Task<IActionResult> Edit(long Id, ProductModel product)
         {
             ViewBag.Categories = new SelectList(_dataContext.Categories, "Id", "Name", product.CategoryId);
             ViewBag.Brands = new SelectList(_dataContext.Brands, "Id", "Name", product.BrandId);
@@ -145,7 +145,7 @@ namespace dotnet_project.Areas.Admin.Controllers
             return View(product);
         }
 
-        public async Task<IActionResult> Delete(int Id)
+        public async Task<IActionResult> Delete(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             if (!string.Equals(product.Image, "noname.jpg"))
